libcshim: sign-extend signed -> wider unsigned conversions (#28)
C converts a signed integer to a wider unsigned type by value modulo
2^N, i.e. it sign-extends first: (opus_uint32)(opus_int8)-4 is
0xFFFFFFFC. The shim zero-extended instead - uint32(uint8(v)) - in
Uint32FromInt8, Uint32FromInt16, Uint64FromInt16 and Uint64FromInt32,
and UintptrFromInt32 did the same, turning a negative pointer offset
into roughly +4 GiB on 64-bit targets.

The SILK encoder hits this constantly: silk_NLSF_residual_dequant
computes (opus_int32)((opus_uint32)indices[i] << 10) on small signed
quantization indices, so any negative index became a large positive
residual, and the encoder rebuilt a nonsense LPC spectrum for nearly
every frame. SILK and hybrid packets were garbage - decoded speech had
an SNR around -19 dB against its input (the error louder than the
signal), at 2-3x the requested VBR bitrate. CELT barely uses these
conversions, which is why music and CELT-only speech were fine.

Checked against libopus 1.6.1 built from source as plain C
(--disable-intrinsics): with this change, opus-go's SILK, hybrid, VoIP
and restricted-SILK output is bit-identical to it, packet for packet, on
real TTS speech clips. Before it, the streams diverged from the first
frame, and a stage-by-stage probe of silk_encode_frame_FLP put the first
difference in silk_find_pred_coefs_FLP's quantized NLSFs.

Tests: the conversions themselves, and the encoder's copy of
silk_NLSF_decode against the decoder's - 2,999 of 3,000 random index
vectors disagreed before this change.

+++ libcshim/libcshim.go
@@ -360,17 +360,17 @@ func Int64FromInt32(v int32) int64   { return int64(v) }
 
 func Uint16FromInt32(v int32) uint16   { return uint16(v) }
 func Uint16FromInt16(v int16) uint16   { return uint16(v) }
-func Uint32FromInt16(v int16) uint32   { return uint32(uint16(v)) }
+func Uint32FromInt16(v int16) uint32   { return uint32(v) }
 func Uint32FromInt32(v int32) uint32   { return uint32(v) }
-func Uint32FromInt8(v int8) uint32     { return uint32(uint8(v)) }
+func Uint32FromInt8(v int8) uint32     { return uint32(v) }
 func Uint32FromUint32(v uint32) uint32 { return v }
-func Uint64FromInt16(v int16) uint64   { return uint64(uint16(v)) }
-func Uint64FromInt32(v int32) uint64   { return uint64(uint32(v)) }
+func Uint64FromInt16(v int16) uint64   { return uint64(v) }
+func Uint64FromInt32(v int32) uint64   { return uint64(v) }
 func Uint64FromInt64(v int64) uint64   { return uint64(v) }
 func Uint64FromUint64(v uint64) uint64 { return v }
 func Uint8FromInt16(v int16) uint8     { return uint8(uint16(v)) }
 func Uint8FromInt32(v int32) uint8     { return uint8(v) }
-func UintptrFromInt32(v int32) uintptr { return uintptr(uint32(v)) }
+func UintptrFromInt32(v int32) uintptr { return uintptr(v) }
 
 // GoString reads a NUL-terminated C string from p.
 func GoString(p uintptr) string {
